Magic casting and experience movement
Added an experience cube that will follow the player when they enter within a certain radius of it and once picked up gives 1xp. Made it so shooting with right click now costs mana. Made mana rechargeable in the same way as stamina.

+++ Assets/Scripts/UI/ManaBar.cs
@@ -1,3 +1,4 @@
+using System.Collections;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -5,12 +6,12 @@ public class ManaBar : MonoBehaviour
 {
     public Slider manaSlider;
 
-    public void SetMaxMana(int mana)
+    public void SetMaxMana(float mana)
     {
         manaSlider.maxValue = mana;
         manaSlider.value = mana;
     }
-    public void SetMana(int mana)
+    public void SetMana(float mana)
     {
         manaSlider.value = mana;
     }
